Clean-up of `filter` implementation (#614)
