+++ pyam/core.py
@@ -11,7 +11,7 @@ import pandas as pd
 from pathlib import Path
 from tempfile import TemporaryDirectory
 
-from pyam.filter import filter_by_time_domain
+from pyam.filter import filter_by_time_domain, filter_by_year, filter_by_dt_arg
 
 try:
     from datapackage import Package
@@ -38,10 +38,6 @@ from pyam.utils import (
     merge_meta,
     find_depth,
     pattern_match,
-    years_match,
-    day_match,
-    datetime_match,
-    time_match,
     isstr,
     islistable,
     print_list,
@@ -51,7 +47,9 @@ from pyam.utils import (
     IAMC_IDX,
     SORT_IDX,
     ILLEGAL_COLS,
-    FILTER_DATETIME_ATTRS,
+)
+from pyam.filter import (
+    datetime_match,
 )
 from pyam.read_ixmp import read_ix
 from pyam.plotting import PlotAccessor
@@ -75,7 +73,7 @@ from pyam.index import (
 )
 from pyam.time import swap_time_for_year, swap_year_for_time
 from pyam._debiasing import _compute_bias
-from pyam.logging import deprecation_warning, raise_data_error
+from pyam.logging import raise_data_error
 
 logger = logging.getLogger(__name__)
 
@@ -1785,45 +1783,14 @@ class IamDataFrame(object):
 
             elif col == "year":
                 levels, codes = get_index_levels_codes(self._data, self.time_col)
-                if self.time_col == "time":
-                    levels = [
-                        i.year if isinstance(i, pd.Timestamp) else i for i in levels
-                    ]
-                matches = years_match(levels, values)
-                keep_col = get_keep_col(codes, matches)
+                keep_col = filter_by_year(self.time_col, values, levels, codes)
 
-            elif col in ["month", "hour"]:
+            elif col in ["month", "hour", "day"]:
                 if self.time_col != "time":
                     logger.error(f"Filter by `{col}` not supported with yearly data.")
                     return np.zeros(len(self), dtype=bool)
 
-                def time_col(x, col):
-                    return getattr(x, col) if isinstance(x, pd.Timestamp) else None
-
-                data = self.get_data_column("time").apply(lambda x: time_col(x, col))
-                if col in FILTER_DATETIME_ATTRS:
-                    keep_col = time_match(data, values, *FILTER_DATETIME_ATTRS[col])
-                else:
-                    keep_col = np.isin(data, values)
-
-            elif col == "day":
-                if self.time_col != "time":
-                    logger.error(f"Filter by `{col}` not supported with yearly data.")
-                    return np.zeros(len(self), dtype=bool)
-
-                if isinstance(values, str):
-                    wday = True
-                elif isinstance(values, list) and isinstance(values[0], str):
-                    wday = True
-                else:
-                    wday = False
-
-                if wday:
-                    days = self.get_data_column("time").apply(lambda x: x.weekday())
-                else:  # ints or list of ints
-                    days = self.get_data_column("time").apply(lambda x: x.day)
-
-                keep_col = day_match(days, values)
+                keep_col = filter_by_dt_arg(col, values, self.get_data_column("time"))
 
             elif col == "time":
                 if self.time_col != "time":
@@ -1843,7 +1810,6 @@ class IamDataFrame(object):
                     has_nan=True,
                     return_codes=True,
                 )
-
                 keep_col = get_keep_col(codes, matches)
 
             else:
